Add files via upload
This Project is Bike Buyers Dataset for Exploratory Data Analysis. The dataset which is used in this project is taken from Kaggle and has details of 1000 users from different backgrounds and whether or not they buy a bike. This project is done by Creating pivot tables, Dashboard visualization on Microsoft Excel.